+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2022-12-05
+### Added
+- [Service Catalog AppRegistry](https://docs.aws.amazon.com/servicecatalog/latest/arguide/intro-app-registry.html) 
+resource to register the CloudFormation template and underlying resources as an application in both Service Catalog 
+AppRegistry and AWS Systems Manager Application Manager.
+### Updated
+- CDK version 1.173.0
+
 ## [1.4.1] - 2021-08-23
 ### Fixed
 - Upgraded Python package versions